Move worker channel connection utility
